+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+canary:
+
+  - now automatically infer the port from the protocol
+  - start protocols' proxy servers
+  - forward request to proxy servers
+
 0.0.1:
+
   - basic engine (not yet working)